tickets/src/providers/StreamingProvider/implementations/listeners/OrderCancelledListener.ts
import { Listener, IOrderCancelledEvent, Subjects } from '@ictickets/common';
import { Ticket } from '@models/Ticket';
import { Message } from 'node-nats-streaming';
import { TicketUpdatedPublisher } from '../publishers/TicketUpdatedPublisher';
import { queueGroupName } from './QueueGroupName';

export class OrderCancelledListener extends Listener<IOrderCancelledEvent> {
  readonly subject = Subjects.OrderCancelled;

  queueGroupName = queueGroupName;

  async onMessage(data: IOrderCancelledEvent['data'], msg: Message) {
    const ticket = await Ticket.findById(data.ticket.id);

    if (!ticket) {
      throw new Error('Ticket not found');
    }

    ticket.set({ orderId: undefined });

    await ticket.save();
    await new TicketUpdatedPublisher(this.client).publish({
      id: ticket.id,
      version: ticket.version,
      title: ticket.title,
      price: ticket.price,
      userId: ticket.userId,
      orderId: ticket.orderId,
    });

    msg.ack();
  }
}
